% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_pu.R
\name{Rmd_bind}
\alias{Rmd_bind}
\title{Concatena Capítulos De Um Power Up}
\usage{
Rmd_bind(dir = ".", titulo, output_file = NULL)
}
\description{
Concatena os capítulos de um power up em um único arquivo index.html.
}

